browser-agent/src/requirements.txt
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
betterpath==0.2.2
certifi==2025.1.31
charset-normalizer==3.4.1
chromedriver-autoinstaller==0.6.4
click==8.1.8
distro==1.9.0
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
l==0.11.0
langchain==0.3.14
langchain-anthropic==0.3.1
langchain-core==0.3.43
langchain-openai==0.3.0
langchain-text-splitters==0.3.6
langgraph==0.3.5
langgraph-checkpoint==2.0.18
langgraph-prebuilt==0.1.2
langgraph-sdk==0.1.55
langsmith==0.2.11
msgpack==1.1.0
multidict==6.1.0
numpy==2.2.3
openai==1.59.7
orjson==3.10.15
outcome==1.3.0.post0
packaging==24.2
playwright==1.50.0
propcache==0.3.0
pydantic==2.10.6
pydantic_core==2.27.2
pyee==12.1.1
PySocks==1.7.1
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
selenium==4.8.0
setuptools==75.9.1
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.38
tenacity==9.0.0
tiktoken==0.9.0
tqdm==4.67.1
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.12.2
urllib3==1.26.20
vcversioner==2.16.0.0
wsproto==1.2.0
yarl==1.18.3
zope.interface==7.2
zstandard==0.23.0
